+++ docs/CHANGES.rst
@@ -6,6 +6,9 @@ Changes
 
 1.11.0 (unreleased)
 -------------------
+- Add removeItems utils and listSolicitOpinionsTo method
+  [fngaha]
+
 - Add getFirstDeposit and _getFirstEvent method
   [fngaha]